requirements.txt
numpy
pandas
swig
gymnasium
ccxt
stable-baselines3
ta
python-dotenv
optuna
torch
tensorboard
scipy